Script to process the NIST handwritten dataset.
The script generates a CSV file from the NIST by_class dataset of handwritten images. To use, unzip the dataset downloaded from https://www.nist.gov/srd/nist-special-database-19 to a folder nist_by_class and run the script. The png images are cropped to 64x64 pixels. The first column has the ASCII value of that character, and columns 2 through 4097 contain the bits of the flattened image.